requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.10.1
django-heroku==0.0.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
Unipath==1.1
whitenoise==5.3.0
pycodestyle==2.7.0
PyYAML==6.0
uritemplate==4.1.1
django-rest-swagger==2.2.0
coreapi==2.3.3
drf-yasg==1.20.0
drf-generators==0.5.0


